Still working on removing only local folders

+++ nyx.py
@@ -88,28 +88,30 @@ class order:
 	def add(self, orderNumber, server, directory):
 		SQL = "INSERT INTO orders (ordernumber, status, server,directory) VALUES (%s,%s,%s,%s);" # Note: no quotes
 		data = (orderNumber, "NEW", server, directory)
-		s = sql()
-		r = s.insert(SQL,data)
+		sql_c = sql()
+		r = sql_c.insert(SQL,data)
 		return r
 
 	def remove(self,o):
-		s = sql()
+		sql_c = sql()
+		utils_c = utils()
 		SQL = "SELECT * FROM deleteorder WHERE ordernumber = %s"
 		data = (o,)
-		rows = s.select(SQL,data)
+		rows = sql_c.select(SQL,data)
 		for row in rows:
 			orderNumber = row[0]
 			notice = row[1]
 			status = row[2]
 			directory = row[3]
 			folder = os.path.join(directory,str(orderNumber))
-			print('Order', orderNumber,'(',notice, ') has the status', status)
+			print('Order {order} [{notice}] has the status {status}'.format(order = orderNumber, notice = notice, status = status)
 			question = 'Are you sure you want to delete this order at {d} ?'.format(d=folder)
 			decision = query_yes_no(question,  default="yes")
 			if decision == 'yes' and os.path.exists(folder):
-				self.deletefiles(folder)
-				self.deletefolder(folder)
-				s.setOrderStatus(row[0],'DELETED')
+				utils_c.deletefiles(folder)
+				utils_c.deletefolder(folder)
+				if not os.path.exists(folder):
+					sql_c.setOrderStatus(orderNumber,'DELETED')
 			else:
 				print('Nothing to delete.')
 		exit()
@@ -354,12 +356,18 @@ class sql:
 
 class utils:
 	def deletefiles(self,dir):
-		filelist = [ f for f in os.listdir(dir) ]
-		for f in filelist:
-			os.remove(os.path.join(dir, f))
+		if os.path.exists(dir):
+			filelist = [ f for f in os.listdir(dir) ]
+			for f in filelist:
+				os.remove(os.path.join(dir, f))
+		else:
+			print('ERROR: {d} is not a local path'.format(d = dir))
 
 	def deletefolder(self,dir):
-		os.rmdir(dir)
+		if os.path.exists(dir):
+			os.rmdir(dir)
+		else:
+			print('ERROR: {d} is not a local path'.format(d = dir))
 
 	def query_yes_no(self,question, default='yes'):
 	    """Ask a yes/no question via raw_input() and return their answer.
@@ -547,7 +555,7 @@ def main(argv):
 		order_c.remove(orderNumber)
 	elif mode == 'generateFootprint':
 		#proc.footprint.info()
-		proc.footprint.generate()
+		#proc.footprint.generate()
 		#proc.footprint.loadgeomtopgsql()
 
 	exit()
